Moved Introduction class as private static inner class of BlackJackApp
