+++ src/com/javaninjas/blackjack/controller/BlackJackApp.java
@@ -3,7 +3,6 @@ package com.javaninjas.blackjack.controller;
 import com.apps.util.Console;
 import com.apps.util.Prompter;
 import com.javaninjas.blackjack.service.Dealer;
-import com.javaninjas.blackjack.service.Introduction;
 import com.javaninjas.blackjack.service.Player;
 import java.io.IOException;
 import java.nio.file.Files;
@@ -312,4 +311,118 @@ public class BlackJackApp {
         return dealer;
     }
 
+    /**
+     * Introduction Class for BlackJack game. A fun introduction screen based on the 80's movie "Wargames". Provides user
+     * input through a Prompter class written by Jay Rostosky.
+     *
+     * @author Alan Pottinger
+     * @version 1.0
+     */
+    private static class Introduction {
+
+        private Introduction(){
+        }
+        /**
+         * "Wargames" style introduction for blackjack game. Uses a Prompter to prompt client if they would like to see menu
+         * or skip to blackjack game.
+         */
+        public static void introduction() throws InterruptedException {
+            Console.clear();
+            System.out.println("\n \u001b[32m \n" + "GREETINGS PROFESSOR FALKEN.\n");
+            TimeUnit.SECONDS.sleep(3);
+            String prompt = prompter.prompt("\nSHALL WE PLAY A GAME? [Y OR N]\n", "y|Y|n|N", "PLEASE SELECT Y OR N. ");
+            if ("y".equalsIgnoreCase(prompt)) {
+                Console.clear();
+                showWopr();
+                showMenu();
+                String game = prompter.prompt("\nPLEASE SELECT YOUR GAME [1-9]\n", "[1-9]", "INVALID SELECTION");
+                switch (game) {
+                    case "1":
+                        System.out.println("ENJOY YOUR GAME!");
+                        TimeUnit.SECONDS.sleep(3);
+                        break;
+                    case "2":
+                        System.out.println("ACCESS DENIED DUE TO COPYRIGHT INFRINGEMENT");
+                        TimeUnit.SECONDS.sleep(3);
+                        System.out.println("WOULDN'T YOU PREFER A GOOD GAME OF BLACKJACK?");
+                        TimeUnit.SECONDS.sleep(4);
+                        break;
+                    case "3":
+                        System.out.println("PEW PEW...EJECT! EJECT!...CRASH! THE END");
+                        TimeUnit.SECONDS.sleep(3);
+                        System.out.println("WOULDN'T YOU PREFER A GOOD GAME OF BLACKJACK?");
+                        TimeUnit.SECONDS.sleep(4);
+                        break;
+                    case "4":
+                        System.out.println("ISN'T THIS JUST HIDE AND SEEK WITH GUNS?");
+                        TimeUnit.SECONDS.sleep(3);
+                        System.out.println("WOULDN'T YOU PREFER A GOOD GAME OF BLACKJACK?");
+                        TimeUnit.SECONDS.sleep(4);
+                        break;
+                    case "5":
+                        System.out.println("SYSTEM ERROR: TOO MUCH SAND WHERE IT DOESN'T BELONG!");
+                        TimeUnit.SECONDS.sleep(3);
+                        System.out.println("WOULDN'T YOU PREFER A GOOD GAME OF BLACKJACK?");
+                        TimeUnit.SECONDS.sleep(4);
+                        break;
+                    case "6":
+                        System.out.println("THIS GAME HAS BEEN REPLACED WITH DRONES");
+                        TimeUnit.SECONDS.sleep(3);
+                        System.out.println("WOULDN'T YOU PREFER A GOOD GAME OF BLACKJACK?");
+                        TimeUnit.SECONDS.sleep(4);
+                        break;
+                    case "7":
+                        System.out.println("THIS GAME HAS BEEN DELETED DUE TO GRAPHIC CONTENT");
+                        TimeUnit.SECONDS.sleep(3);
+                        System.out.println("WOULDN'T YOU PREFER A GOOD GAME OF BLACKJACK?");
+                        TimeUnit.SECONDS.sleep(4);
+                        break;
+                    case "8":
+                        System.out.println("COUGH...COUGH...THE END");
+                        TimeUnit.SECONDS.sleep(3);
+                        System.out.println("WOULDN'T YOU PREFER A GOOD GAME OF BLACKJACK?");
+                        TimeUnit.SECONDS.sleep(4);
+                        break;
+                    case "9":
+                        System.out.println("EVERYONE IS DEAD...ARE YOU HAPPY NOW?");
+                        TimeUnit.SECONDS.sleep(3);
+                        System.out.println("WOULDN'T YOU PREFER A GOOD GAME OF BLACKJACK?");
+                        TimeUnit.SECONDS.sleep(4);
+                }
+            } else {
+                System.out.println("LOGGING OFF: W.O.P.R");
+                TimeUnit.SECONDS.sleep(4);
+
+            }
+        }
+
+        /**
+         * Retrieves the wopr ascii art file and displays it.
+         */
+        private static void showWopr() {
+            System.out.println("\n\n");
+            try {
+                Files.lines(Path.of("resources", "wopr.txt")).forEach(System.out::println);
+            } catch (IOException e) {
+                e.printStackTrace();
+            }
+        }
+
+        /**
+         * Displays "Wargames" style menu.
+         */
+        private static void showMenu() {
+            System.out.println("\nLOGON: LIST GAMES");
+            System.out.println("\n\n" +
+                    "1. BLACKJACK\n\n" +
+                    "2. FALKEN'S MAZE\n\n" +
+                    "3. FIGHTER COMBAT\n\n" +
+                    "4. GUERRILLA ENGAGEMENT\n\n" +
+                    "5. DESERT WARFARE\n\n" +
+                    "6. AIR-TO-GROUND ACTIONS\n\n" +
+                    "7. THEATERWIDE TACTICAL WARFARE\n\n" +
+                    "8. THEATERWIDE BIOTOXIC AND CHEMICAL WARFARE\n\n" +
+                    "9. GLOBAL THERMONUCLEAR WAR\n\n");
+        }
+    }
 }
